imports organized into src.common.consts and src.common.interfaces
