+++ src/cl/branch_fu.py
@@ -1,10 +1,13 @@
 # Branch calculation functional unit.
 # Responsible for redirecting the pipeline when a branch is mispredicted.
 
-from pymtl3 import Component, InPort, OutPort, update
+from pymtl3 import Component, InPort, OutPort, update, bitstruct, mk_bits, clog2
 from pymtl3.stdlib.ifcs import RecvIfcRTL, SendIfcRTL
-from src.cl.decode import (
-    MicroOp,
+from src.common.interfaces import (
+    BranchUpdate,
+    MicroOp
+)
+from src.common.consts import (
     BRANCH_FUNCT_UNIT,
     BFU_BEQ,
     BFU_BNE,
@@ -23,7 +26,7 @@ class BranchFU( Component ):
         s.uop = InPort(MicroOp)
 
         # TODO: redirect address for now, will contain more metedata later
-        s.mispredict = SendIfcRTL(32)
+        s.br_update = SendIfcRTL(BranchUpdate)
 
         @update
         def updt():
@@ -31,6 +34,8 @@ class BranchFU( Component ):
             rs1_neg = s.rs1_din[32-1]
             rs2_neg = s.rs2_din[32-1]
             sign_diff = rs1_neg ^ rs2_neg
+            s.br_update.en @= (s.uop.funct_unit == BRANCH_FUNCT_UNIT)
+            s.br_update.msg.tag @= s.uop.br_tag
 
             # if branch taken...
             if (((s.uop.funct_op == BFU_BEQ) & (s.rs1_din == s.rs2_din)) |
@@ -43,18 +48,20 @@ class BranchFU( Component ):
                     (~sign_diff & (s.rs1_din >= s.rs2_din)))) |
                 ((s.uop.funct_op == BFU_BLTU) & (s.rs1_din < s.rs2_din)) |
                 ((s.uop.funct_op == BFU_BGEU) & (s.rs1_din >= s.rs2_din))):
+
                 # mispredict if branch is predicted not taken but taken
-                s.mispredict.en @= \
-                    (s.uop.funct_unit == BRANCH_FUNCT_UNIT) & ~s.uop.branch_taken
-                s.mispredict.msg @= s.uop.pc + s.uop.imm
+                s.br_update.msg.mispredict @= ~s.uop.branch_taken
+                s.br_update.msg.target @= s.uop.pc + s.uop.imm
 
             # if branch not taken...
             else:
                 # mispredict if branch is predicted taken but not taken
-                s.mispredict.en @= \
-                    (s.uop.funct_unit == BRANCH_FUNCT_UNIT) & s.uop.branch_taken
-                s.mispredict.msg @= s.uop.pc + 8
-
+                s.br_update.msg.mispredict @= s.uop.branch_taken
+                s.br_update.msg.target @= s.uop.pc + 8
 
     def line_trace( s ):
-        return (f"rs1: {s.rs1_din} rs2: {s.rs2_din} mispredict: {s.mispredict.msg if s.mispredict.en else 'none'}")
+        en = s.br_update.en
+        return (f"rs1: {s.rs1_din} rs2: {s.rs2_din} "
+                f"mispredict: {s.br_update.msg.mispredict if en else 'none'} "
+                f"target: {s.br_update.msg.target if en else 'none'} "
+                f"tag: {s.br_update.msg.tag if en else 'none'}")
