+++ vcd/test_dispatch.vcd
@@ -1,5 +1,5 @@
 $date
-  Sun Aug 28 17:02:53 2022
+  Tue Aug 30 11:53:47 2022
 $end
 $version
   PyMTL 3 (Mamba)
@@ -10,32 +10,32 @@ $end
 
 $scope module top $end
   $var reg 336 + to_mem_issue $end
+  $var reg 5 , rob_idx $end
   $var reg 1 ! clk $end
-  $var reg 336 , in_ $end
-  $var reg 1 & reset $end
-  $var reg 336 - to_rob $end
+  $var reg 1 ( reset $end
+  $var reg 4 - mem_q_tail $end
   $var reg 336 . to_int_issue $end
-  $var reg 4 / mem_q_tail $end
-  $var reg 5 0 rob_idx $end
+  $var reg 336 / in_ $end
+  $var reg 336 0 to_rob $end
   $scope module uop2_dispatch $end
-    $var reg 1 & reset $end
-    $var reg 168 ' to_int_issue $end
-    $var reg 1 ! clk $end
-    $var reg 168 * to_mem_issue $end
+    $var reg 168 " to_mem_issue $end
     $var reg 168 # to_rob $end
-    $var reg 4 1 mem_q_idx $end
-    $var reg 5 2 rob_idx $end
     $var reg 168 $ in_ $end
+    $var reg 1 ( reset $end
+    $var reg 168 * to_int_issue $end
+    $var reg 1 ! clk $end
+    $var reg 5 1 rob_idx $end
+    $var reg 4 2 mem_q_idx $end
   $upscope $end
   $scope module uop1_dispatch $end
-    $var reg 168 " to_int_issue $end
-    $var reg 1 ! clk $end
+    $var reg 168 ' to_rob $end
     $var reg 4 3 mem_q_idx $end
-    $var reg 168 % to_rob $end
-    $var reg 168 ( in_ $end
+    $var reg 1 ( reset $end
+    $var reg 1 ! clk $end
+    $var reg 168 ) in_ $end
+    $var reg 168 % to_mem_issue $end
+    $var reg 168 & to_int_issue $end
     $var reg 5 4 rob_idx $end
-    $var reg 168 ) to_mem_issue $end
-    $var reg 1 & reset $end
   $upscope $end
 $upscope $end
 $enddefinitions $end
@@ -45,19 +45,19 @@ b0b00000000000000000000000000000000000000000000000000000000000000000000000000000
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 #
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 $
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 %
-b0b0 &
+b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 &
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 '
-b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 (
+b0b0 (
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 )
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 *
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 +
-b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 ,
-b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 -
+b0b00000 ,
+b0b0000 -
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 .
-b0b0000 /
-b0b00000 0
-b0b0000 1
-b0b00000 2
+b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 /
+b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 0
+b0b00000 1
+b0b0000 2
 b0b0000 3
 b0b00000 4
 
@@ -65,14 +65,14 @@ b0b00000 4
 b0b1 !
 
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 "
-b0b1 &
-b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 '
+b0b1 (
+b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 )
 b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 +
-b0b0000 /
-b0b00000 0
-b0b0000 1
-b0b00001 2
-b0b0000 3
+b0b00000 ,
+b0b0000 -
+b0b000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 .
+b0b00001 1
+b0b0000 2
 b0b00000 4
 
 #50
