+++ vcd/test_front_end.vcd
@@ -1,5 +1,5 @@
 $date
-  Sun Aug 28 17:02:53 2022
+  Tue Aug 30 11:53:47 2022
 $end
 $version
   PyMTL 3 (Mamba)
@@ -9,23 +9,23 @@ $timescale
 $end
 
 $scope module top $end
-  $var reg 98 $ fetch_packet $end
-  $var reg 32 % mispredict.msg $end
-  $var reg 32 & pc_next $end
-  $var reg 1 # mispredict.rdy $end
-  $var reg 1 ' mispredict.en $end
-  $var reg 64 ( icache_data $end
-  $var reg 1 " clk $end
+  $var reg 1 $ br_update.en $end
   $var reg 1 ! reset $end
-  $var reg 32 ) pc $end
+  $var reg 32 % pc_next $end
+  $var reg 1 " clk $end
+  $var reg 32 & pc $end
+  $var reg 1 # br_update.rdy $end
+  $var reg 98 ' fetch_packet $end
+  $var reg 36 ( br_update.msg $end
+  $var reg 64 ) icache_data $end
   $scope module icache $end
-    $var reg 1 " clk $end
     $var reg 64 * data_out $end
     $var reg 8 + addr $end
+    $var reg 1 " clk $end
     $var reg 1 ! reset $end
     $scope module memory $end
-      $var reg 1 ! reset $end
       $var reg 1 " clk $end
+      $var reg 1 ! reset $end
     $upscope $end
   $upscope $end
 $upscope $end
@@ -34,12 +34,12 @@ $enddefinitions $end
 b0b0 !
 b0b0 "
 b0b0 #
-b0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 $
+b0b0 $
 b0b00000000000000000000000000000000 %
 b0b00000000000000000000000000000000 &
-b0b0 '
-b0b0000000000000000000000000000000000000000000000000000000000000000 (
-b0b00000000000000000000000000000000 )
+b0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000 '
+b0b000000000000000000000000000000000000 (
+b0b0000000000000000000000000000000000000000000000000000000000000000 )
 b0b0000000000000000000000000000000000000000000000000000000000000000 *
 b0b00000000 +
 
@@ -48,13 +48,10 @@ b0b1 "
 
 b0b1 !
 b0b1 #
-b0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001 $
-b0b00000000000000000000000000000000 %
-b0b00000000000000000000000000001000 &
-b0b0 '
-b0b0000000000000000000000000000000000000000000000000000000000000000 (
-b0b00000000000000000000000000000000 )
-b0b0000000000000000000000000000000000000000000000000000000000000000 *
+b0b00000000000000000000000000001000 %
+b0b00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000001 '
+b0b000000000000000000000000000000000000 (
+b0b0000000000000000000000000000000000000000000000000000000000000000 )
 b0b00000000 +
 
 #50
